Rename package and Docker image.
