+++ .github/workflows/publish.yml
@@ -16,7 +16,7 @@ jobs:
         run: |
           docker login --username ${{ secrets.DOCKER_USERNAME }} --password ${{ secrets.DOCKER_PASSWORD }}
 
-          if DOCKER_CLI_EXPERIMENTAL=enabled docker manifest inspect multiversx/build-contract-rust:${GITHUB_REF_NAME} >/dev/null; then
+          if DOCKER_CLI_EXPERIMENTAL=enabled docker manifest inspect multiversx/sdk-rust-contract-builder:${GITHUB_REF_NAME} >/dev/null; then
               echo "Image already published. Will NOT publish."
               exit 1
           else
@@ -42,4 +42,4 @@ jobs:
           push: true
           platforms: linux/amd64,linux/arm64
           file: ./Dockerfile
-          tags: multiversx/build-contract-rust:${{ github.ref_name }}
+          tags: multiversx/sdk-rust-contract-builder:${{ github.ref_name }}
