--- build_contract_rust/__init__.py
@@ -1,3 +1,0 @@
-from build_contract_rust import builder
-
-__all__ = ["builder"]
