+++ multiversx_sdk_rust_contract_builder/__init__.py
@@ -1,0 +1,3 @@
+from multiversx_sdk_rust_contract_builder import builder
+
+__all__ = ["builder"]
