+++ testdata/input/empty/.gitignore
@@ -3,5 +3,5 @@
 /target/
 */target/
 
-# The erdpy output
+# The build output
 output*
